Selling Price Added

+++ ShoeShop/DTOs/ProductViewDto.cs
@@ -20,7 +20,7 @@ namespace ShoeShop.DTOs
         public int Size { get; set; }
         public int Count { get; set; }
         public string BranchName { get; set; }
-        public int SellingPrice { get; set; }
+        public double SellingPrice { get; set; }
         public int TempId { get; set; }
         public int BranchId { get; set; }
         public Branch Branch { get; set; }
